+++ framework/CHANGELOG.md
@@ -9,13 +9,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Install modules on account or Sub-account creation
+- Manager stores his sub-accounts and sub-accounts can register or unregister in case of ownership change
+- Query on module factory to see how much funds needs to be attached for installing modules 
 
 ### Changed
 - Updated fetch_data arguments of CwStakingCommand
 - StakingInfoResponse now returns staking target(which is either contract address or pool id) instead of always staking contract address
+- Owner of the sub-accounts now Proxy, allowing modules to interact with sub-accounts
+- Install modules replaced install module method on module factory to reduce gas consumption for multi-install cases 
 
 ### Fixed
 - Partially fixed cw-staking for Osmosis
+- Manager governance now changes only after new "owner" claimed ownership
 - Fixed and separated cw-staking and dex adapters for kujira
 
 ## [0.17.2] - 2023-07-27
@@ -32,6 +37,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Ability to set admin to native contracts during instantiation
+- Query handler for module data
 - Added neutron
 
 ### Changed